Second commit

+++ Car_Leasing/src/main/java/com/ado/leasing/dao/UserDAOInterface.java
@@ -10,7 +10,7 @@ public interface UserDAOInterface {
 	
 	public void saveUser(User theUser);
 	
-	public User getUser(int id);
+	public User getUser(BigInteger id);
 	
 	public void deleteUser(BigInteger id);
